// src/app/models/municipios_enderecos.ts
export class Municipios_enderecos {
    id:           string = "";
    cidade:           string;
    rua:              string;
    bairro:           string;
    complemento:      string;
    numero:           string;
    cep:              string;
   

    // adicionais
    perfil: string;

/*     constructor() {
        this.cidade = '';
        this.rua = '';
        this.bairro = '';
        this.complemento = '';
        this.numero = '';
        this.cep = '';
            } */
}